joget_utility/requirements.txt
requests>=2.31.0
PyYAML>=6.0
mysql-connector-python>=8.0.0
pandas>=2.1.0